+++ CHANGELOG.md
@@ -15,6 +15,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+### Added
+
+- diagonal and unfold are CPU views
+
 ---
 
 ## [3.13.0] — 2026-09-17